Added accessors for testCount, assertionCount; more assertionTests
- Added tests for string equality assertions
- Improved reliability for failure counting

+++ include/partest/bootstrap.h
@@ -61,12 +61,12 @@ namespace partest
 		testRunner().printAllTestTrees();
 	}
 
-	inline unsigned getTopLevelFailures()
+	inline size_t getTopLevelFailures()
 	{
 		return testRunner().getTopLevelFailures();
 	}
 
-	inline unsigned getAssertionFailureCount()
+	inline size_t getAssertionFailureCount()
 	{
 		return testRunner().getAllAssertionFailures();
 	}
